Efficiency Addition Part 1
Did some things with the engine to allow more efficient running on the
CPU end of things.  Added a bool to the game object class that handles
whether or not an object should check for collisions.  Also added a new
function for adding large amounts of objects to the object handler.
Hand in the index you would like for it to start at and it will start
there.  Example can be seen in the Level1 class.
Known Issues: Loose primitive does not stay with base enemy, will fix.
There is a cap on the amount of objects that can exist before the
collision function takes up to much cpu power.  Currently working to fix
this.