Done! No more lag!
Let's just say I made huge mistake... Like really badly! I accidently
put trigger collisions in a for loop that looped through all objects.
That was not good! That was the cause of most of the lag.  I also,
because of this, saw how to implement a better way to check for
collisions.  Everything is a lot faster now.