Shooting and Other things
I added a destroy function to remove objects from the game.  I also
added shooting, with a base projectile class.  Destroy can not be called
during the collision code so an example of how to  use it on collision
is given in the provided collision code for the projectile.  I also
updated flappy bird game play to be more accurate.