Graphic EdgeDetection
Detects edges and makes graphics as appropriate.

Henry, you'll see the example graphics I already created, and I also
changed the template slightly (defaultEdge).  Corners are not going to
be a thing, they weren't needed much and were giving me a headache.

The only issue left is figuring out how to use layers correctly, as when
you go into a cave you can notice that some of the tiles are going over
some of our edges.  Tried messing around with the Depth overload of
Draw() but to no success.