Collision Code Fixes pt 2
I added a C_NoCollisions function to the basic game object that gets
called if there were no collisions at all.
I also checked out the code that Mike added and it was kind of funky and
didn't seem to serve any purpose at all.  I set up a new baseline for
the collision code to be worked under, and also removed an issue that
Mike was taking advantage of(I don't know how) where the object was
colliding with itself.

The code that mike was designing seemed to try to find and end all be
all solution when colliding with edge tiles.  This is impossible, the
way that i started(and couldn't finish because of dinner) tries to check
for each edge.  I was currently debugging it when I had to stop.  Please
work with the outline I set up because that is the way that the
collisions need to work.  Take a look, i put basic comments in, and I am
aware it does not work.