Started on Mouse
Had to leave so I am just committing with the cursor object created,
haven't had a chance to begin work on it. I also added to the player
class you have to be holding down right to continue. I may change this
when I get back to if they are colliding with the ground and the last
input was the direction, they stop, otherwise they continue, but that's
a maybe for now. It does work fine right now after all. Anyway, I'll get
back to work asap.