Minor Changes
I made a response to Mike's in code question.
Also changed zOrder to something that was not a joke.  I am going to be
harsh about this.  What if we wanted for some reason to hide the mouse
under something?  Or wanted an actual reference to know what zOrder it
was on.  Do not do things like that mike because you want a joke to
happen in a serious code project.