Debug MODE
Check it out! Alt+Tilda(~) to enter into debug mode.  Type help for a
list of commands.  GameObject is case sensitive to the name in the
dictionary.  It is awesome, super duper awesome.  I'll be adding more as
we progress.  Shift is a little jank, left shift to cap, right shift to
lower. Also, Mouse will show what object it is hovering over while on
screen. But the mouse collision position is a little messed up so we
will work on that.