Fall off edges with no natural jitter
Fixed the collisions, re did the way we do collision detection entirely.
Enjoy.  Still a small issue with collision from the left side, and from
the bottom.  I will address those when next I have time.  Enjoy the
newest version and update the player movement and such until it is the
way that you like it.