Skills implemented
Added the ability for skills to be used.  Also enabled double jumping as
well as wings, wings allow you to glide.  Triple jumping can be enabled,
up to you to try out.  Dashing changed to work with the shift key, and q
and e can be used to change your active skill.  If you want to add a new
skill, edit the skill class in the Skills folder.  All of them contain a
player object so you can interact with the player using the skill.  I
made dash a skill to show this off, but some of dashing /had/ to be done
in Player for the sake of simplicity.  This should reduce the number of
potential collisions from working in player in the future.