Temporary collision fix
To preface my attempt at fixing this rather odd collision bug, I could
not run video studio in debug mode. The game would not launch.  There is
an issue where the player's collision with an object from the left
collision state is not happening when we decide to add in the background
tiles.  This is a very, very odd effect and kind of perplexes me.  I
have impolemented a temporary fix to only the player for while I work a
way around this.  Thank you.