Aidan and Mike Fixes and additions
Aidan: I fixed the collisions, the issue had to do with array indexing
that worked when we had fewer objects because it always happened to
refer to the correct thing, this broke when we had more than one thing
and the indexing broke.  This had to do with the draw order of things
because the background tiles are put below the player and therefore the
player was no longer the 0th thing in the array.  I FIXED IT! Yay, sp,
it was because I used the wrong array for one part of the collision
check code that caused it to work most of the time because of all of the
similar dimensions.  Otherwise it is fixed, yay!
As for Mike's work, he is working on adding art and other cool things
like the working buttons.  Pretty awesome stuff.