Buttons and More
Added the new player sprite to content but did not change it.
Additionally buttons are tested and do work, I just need to put the
levels together now. Aidan and I spoke and decided it would be easier
for me to just create all the buttons' textures independently of one
another rather than alter the xml/xnb files of spritefonts to fit what I
want to do. Expect all the levels done by the end of this coming
weekend.