SWORDS!
Well, A sword drawn by me. So not great... ANYWAY,it is pretty neat,
left click to attack.  Working on coming up with the implementation of
some other things.
BUGS: Sometimes the onions get stuck in the walls.
Collision is still broken, will fix when I have a debugger.
Otherwise good I believe.