Chaotic Reset
Added Chaotic Reset, I also modified the location of some of the things
that were included in the Level class to be included in the base state.
Now we can store the current leveltype as a base "Level type" variable,
I also moved the level enum to the base level as well.  Also, a public
static currState that is always accesible.