Player2.0
New Player sprite and some little changes to go along with it. I have to
fix the player sprite still but I'll get that done later (back foot
needs to be lowered by 1 pixel for some frames and the bounding box
needs to the max width of the texture, my bad).