Dialouge Boxes are in!
To make a new dialouge Box event, create a new object that inherits from
event, and then add it to the Dialouge box class' list of events.  Set
the index to the index that you want to match the text file entitled
AllText.  To add more text, open the file, push enter, type @ and a new
index, and then type away.  Do not use newlines in your sentences. They
are what I use to parse the sentences.