Stamina Bars and Health Bars added
Need to be moved to be a separate object, but this will work for now.
Just wanted to get it in quickly.  The drawing code needs to be done on
top of all other code because it's the UI.  Perhaps we should create a
UI object where the drawing code is done.

Also updated Portal dialogue box.